Adding Concurrent Mark timing stats in ConcurrentPhaseStatsBase, and BalancedGC flags
This change introduces a few flags which are used to keep track of Concurrent marking work timing/cost.

- `_concurrentMarkStartTime` keeps track of the cpu time for a concurrent mark increment
- Added additional flags for heap expansion and contraction.

This change also includes a slight modification to the way `dnssExpectedTimeRatio(Maximum|Minimum)` are used, so other GC policies can use them (or set their own defaults) if desired.
The implementation of `MM_UserSpecifiedParameterDouble` mirrors that of `MM_UserSpecifiedParameterUDATA`.
With this, initialization of `dnssExpectedTimeRatio(Maximum|Minimum)` for gencon now occurs in `MM_ConfigurationGenerational::initialize()` - but the logic/behaviour remains unchanged.

Lastly, modified and added a handful of tracepoints.

Signed-off-by: Cedric Hansen <cedric.hansen@ibm.com>